Push AST root and state down to AbstractMQLTranslator
